Move slack section to component
